+++ bot/src/db/models.ts
@@ -1,7 +1,0 @@
-import User from "./models/user.model";
-import Transfer from "./models/transfer.model";
-
-export const models = {
-  User,
-  Transfer,
-};