+++ bot/src/services/config/config.interface.ts
@@ -1,0 +1,4 @@
+export interface ConfigServiceI {
+  get(key: string): string;
+  set(key: string, value: string): void;
+}