+++ server/src/middlewares/validators/auth.validator.ts
@@ -1,6 +1,7 @@
 import { Request, Response, NextFunction } from "express";
 import jwt from "jsonwebtoken";
 import dotenv from "dotenv";
+import logger from "../../utils/logger";
 
 dotenv.config();
 
@@ -9,6 +10,7 @@ class AuthValidator {
     if (req.method === "OPTIONS") next();
     const token = req.headers.authorization?.split(" ")[1];
     if (!token) {
+      logger.error("auth token not provided");
       return res.status(401).json({ message: "Token not provided" });
     }
 
@@ -20,6 +22,7 @@ class AuthValidator {
       req.body.__user = decoded;
       next();
     } catch (err) {
+      logger.error("auth invalid token");
       return res.status(403).json({ message: "Invalid token" });
     }
   }