+++ server/src/services/config/config.interface.ts
@@ -1,0 +1,3 @@
+export interface ConfigServiceI {
+  get(key: string): string;
+}